1391 - escape character before an apostrophe in date formats are no longer removed by EPPlus
